Internally organize lane-related types and code.
